Balanced Spawning
● each enemy and powerup
now has its own spawn chance, which can be set on the object

This chance depends on the random number between 0 and the sum of all of the chances of specific objects. Lets say powerups.
Random between 0 and powerupsChances sum gives us a float, by which we can determine which powerup to summon.